refactor: apply review(3) edit tests
